agregado 3 reportes pdf en la seccion de descarga, quitado los botones eliminar, cambiado algunos permisos de admin y secreteari, agregado el plugin de calendario, mejorados los select con buscadores, y algunos detalles mas

+++ resources/views/asignaciones/index.blade.php
@@ -29,44 +29,48 @@
                   <table class="table table-striped table-bordered table-hover" id="dataTables-example">
                     <thead>
                       <tr>
+                        <th>Fecha y Hora</th>
                         <th>Autobus</th>
                         <th>Ruta</th>
-                        <th>Fecha y Hora</th>
-                        <th>Acciones</th>
+                        @if (Auth::user()->admin())
+                          <th>Acciones</th>
+                        @endif
                       </tr>
                     </thead>
                     <tbody>
                       @foreach($rutas as $ruta)
                         <tr class="">
-                          <td>{{ $ruta->auto->numero }} - {{ $ruta->auto->matricula }}</td>
-                          <td>{{ $ruta->ruta->nombre }}</td>
                           <td>{{ $ruta->created_at }}</td>
-                          <td>
-                            <div class="" role="">
+                          <td>{{ $ruta->auto->matricula }} - {{ $ruta->auto->numero }}</td>
+                          <td>{{ $ruta->ruta->nombre }}</td>
+                          @if (Auth::user()->admin())
+                            <td>
+                              <div class="" role="">
 
-                              <button type="button" class="btn btn-danger btn-xs" data-toggle="modal" data-target="#del{!!$ruta->id!!}"><i class="fa fa-trash"></i> Eliminar</button>
-                              <!-- Modal -->
-                              <div class="modal fade" id="del{!!$ruta->id!!}" tabindex="-1" role="dialog" aria-labelledby="myModalLabel">
-                                <div class="modal-dialog" role="document">
-                                  <div class="modal-content">
-                                    <div class="modal-header">
-                                      <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button>
-                                      <h4 class="modal-title" id="myModalLabel">Desea Eliminar?</h4>
-                                    </div>
-                                    <div class="modal-body">
-                                      Se eliminará de forma permanente
-                                    </div>
-                                    <div class="modal-footer">
-                                      <div class="btn-group">
-                                        <button type="button" class="btn btn-success" data-dismiss="modal">Cancelar</button>
-                                        <a href="asignaciones/{{$ruta->id}}/delete" class="btn btn-danger">Eliminar <i class="icon-trash"></i></a>
+                                <button type="button" class="btn btn-danger btn-xs" data-toggle="modal" data-target="#del{!!$ruta->id!!}"><i class="fa fa-trash"></i> Eliminar</button>
+                                <!-- Modal -->
+                                <div class="modal fade" id="del{!!$ruta->id!!}" tabindex="-1" role="dialog" aria-labelledby="myModalLabel">
+                                  <div class="modal-dialog" role="document">
+                                    <div class="modal-content">
+                                      <div class="modal-header">
+                                        <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button>
+                                        <h4 class="modal-title" id="myModalLabel">Desea Eliminar?</h4>
+                                      </div>
+                                      <div class="modal-body">
+                                        Se eliminará de forma permanente
+                                      </div>
+                                      <div class="modal-footer">
+                                        <div class="btn-group">
+                                          <button type="button" class="btn btn-success" data-dismiss="modal">Cancelar</button>
+                                          <a href="asignaciones/{{$ruta->id}}/delete" class="btn btn-danger">Eliminar <i class="icon-trash"></i></a>
+                                        </div>
                                       </div>
                                     </div>
                                   </div>
-                                </div>
-                              </div> <!-- / Modal -->
-                            </div>
-                          </td>
+                                </div> <!-- / Modal -->
+                              </div>
+                            </td>
+                          @endif
                         </tr>
                       @endforeach()
                     </tbody>
